portfolio_architect/requirements.txt
strands-agents
boto3
bedrock-agentcore
yfinance
pandas